added an endpoint for evaluating goals

+++ KPICatalog.Domain/Dtos/Entities/TypicalGoalInBonusSchemeDto.cs
@@ -5,15 +5,15 @@ public class TypicalGoalInBonusSchemeDto : BaseEntityDto
     public int TypicalGoalId { get; set; }
     public int ParentBSTypicalGoalId { get; set; }
     public int PeriodId { get; set; }
-    public int Weight { get; set; }
-    public int Plan { get; set; }
+    public decimal Weight { get; set; }
+    public decimal Plan { get; set; }
     public int TypeKeyResultId { get; set; }
     public int EvaluationProvider { get; set; }
     public int BonusSchemeLinkMethodId { get; set; }
     public int EvaluationMethodId { get; set; }
     public int RatingScaleId { get; set; }
-    public int? Fact { get; set; }
-    public int? Evaluation { get; set; }
+    public decimal? Fact { get; set; }
+    public decimal? Evaluation { get; set; }
 
     public PeriodDto Period { get; set; }
 }
